# server: improve chained interceptors performance