This improves performance of chained interceptors by removing need to call `getChain(Unary|Stream)Handler` each request. It affects only servers with more than one interceptor.

Here's small benchmark and its result:

```
benchmark                              old ns/op     new ns/op     delta
BenchmarkChainUnaryInterceptor-12      372           209           -43.97%
BenchmarkChainStreamInterceptor-12     404           271           -32.98%

benchmark                              old allocs     new allocs     delta
BenchmarkChainUnaryInterceptor-12      4              4              +0.00%
BenchmarkChainStreamInterceptor-12     4              4              +0.00%

benchmark                              old bytes     new bytes     delta
BenchmarkChainUnaryInterceptor-12      256           128           -50.00%
BenchmarkChainStreamInterceptor-12     256           192           -25.00%
```

```go
func BenchmarkChainUnaryInterceptor(b *testing.B) {
	s := NewServer(ChainUnaryInterceptor(
		func(
			ctx context.Context, req interface{}, info *UnaryServerInfo, handler UnaryHandler,
		) (interface{}, error) {
			return handler(ctx, req)
		},
		func(
			ctx context.Context, req interface{}, info *UnaryServerInfo, handler UnaryHandler,
		) (interface{}, error) {
			return handler(ctx, req)
		},
		func(
			ctx context.Context, req interface{}, info *UnaryServerInfo, handler UnaryHandler,
		) (interface{}, error) {
			return handler(ctx, req)
		},
		func(
			ctx context.Context, req interface{}, info *UnaryServerInfo, handler UnaryHandler,
		) (interface{}, error) {
			return handler(ctx, req)
		},
		func(
			ctx context.Context, req interface{}, info *UnaryServerInfo, handler UnaryHandler,
		) (interface{}, error) {
			return handler(ctx, req)
		},
	))

	b.ResetTimer()
	for i := 0; i < b.N; i++ {
		_, _ = s.opts.unaryInt(context.Background(), nil, nil, func(ctx context.Context, req interface{}) (interface{}, error) {
			return nil, nil
		})
	}
}

func BenchmarkChainStreamInterceptor(b *testing.B) {
	s := NewServer(ChainStreamInterceptor(
		func(srv interface{}, ss ServerStream, info *StreamServerInfo, handler StreamHandler) error {
			return handler(srv, ss)
		},
		func(srv interface{}, ss ServerStream, info *StreamServerInfo, handler StreamHandler) error {
			return handler(srv, ss)
		},
		func(srv interface{}, ss ServerStream, info *StreamServerInfo, handler StreamHandler) error {
			return handler(srv, ss)
		},
		func(srv interface{}, ss ServerStream, info *StreamServerInfo, handler StreamHandler) error {
			return handler(srv, ss)
		},
		func(srv interface{}, ss ServerStream, info *StreamServerInfo, handler StreamHandler) error {
			return handler(srv, ss)
		},
	))

	b.ResetTimer()
	for i := 0; i < b.N; i++ {
		_ = s.opts.streamInt(nil, nil, nil, func(srv interface{}, stream ServerStream) error {
			return nil
		})
	}
}
```

RELEASE NOTES:
- server: improve performance when multiple interceptors are used