@amenzhinsky 2021-06-14
Added one more commit, made it slightly better:

```
benchmark                             old ns/op     new ns/op     delta
BenchmarkChainUnaryInterceptor-12     440           176           -59.95%

benchmark                             old allocs     new allocs     delta
BenchmarkChainUnaryInterceptor-12     4              3              -25.00%

benchmark                             old bytes     new bytes     delta
BenchmarkChainUnaryInterceptor-12     256           80            -68.75%
```