@zasweq 2021-06-14
Can you please add the benchmark to the codebase rather than as a comment here?