@menghanl 2021-06-25
Thanks for the fix!

There's also an interceptor chain on the client side, that seems to have a similar problem. https://github.com/grpc/grpc-go/blob/master/clientconn.go#L372
Are you interested in taking a look at that too? Thanks!